Remove useless includes
