+++ src/dhcp.cpp
@@ -27,7 +27,6 @@
  *
  */
 
-#include <stdexcept>
 #include <cstring>
 #include "endianness.h"
 #include "dhcp.h"
