+++ src/dot1q.cpp
@@ -27,7 +27,6 @@
  *
  */
 
-#include <stdexcept>
 #include <cstring>
 #include "dot1q.h"
 #include "exceptions.h"
