+++ src/icmp.cpp
@@ -27,7 +27,6 @@
  *
  */
 
-#include <stdexcept>
 #include <cstring>
 #ifndef _WIN32
     #include <netinet/in.h>
