+++ src/radiotap.cpp
@@ -32,7 +32,6 @@
 #ifdef TINS_HAVE_DOT11
 
 #include <cstring>
-#include <stdexcept>
 #include "macros.h"
 #ifndef _WIN32
     #if defined(BSD) || defined(__FreeBSD_kernel__)
